Model the output directory for source files generated by annotation processors as a `DirectoryProperty` and replace the convention mapping with a convention on the property.
This change means that JavaCompile tasks added by the Java base plugin for a source set will be up-to-date on first load from the instant execution cache, and will generate source files to the correct location when they do happen to run.

The issue was caused because instant execution (intentionally) ignores convention mappings applied to `Provider` types fields. This issue, and others like it, would have been easier to diagnose if this case were treated as an instant execution serialization problem (and reported, etc).